Factor put interceptor Update tests
