+++ mobile/asa-go/src/components/loginButton.test.tsx
@@ -1,49 +1,57 @@
+import { render, screen, fireEvent } from "@testing-library/react";
+import { Mock, vi } from "vitest";
+import LoginButton from "@/components/LoginButton";
+import { useDispatch } from "react-redux";
 import { authenticate } from "@/slices/authenticationSlice";
-import { createTestStore } from "@/testUtils";
+import { setAxiosRequestInterceptors } from "@/utils/axiosInterceptor";
 import { ThemeProvider, createTheme } from "@mui/material/styles";
-import { fireEvent, render, screen } from "@testing-library/react";
-import { Provider } from "react-redux";
-import { describe, expect, it, vi } from "vitest";
-import LoginButton from "./LoginButton";
 
-// Mock authenticate action
+// Mocks
+vi.mock("react-redux", async () => {
+  const actual = await vi.importActual("react-redux");
+  return {
+    ...actual,
+    useDispatch: vi.fn(),
+  };
+});
+
 vi.mock("@/slices/authenticationSlice", () => ({
   authenticate: vi.fn(() => ({ type: "AUTHENTICATE" })),
-  default: vi.fn(() => null),
 }));
 
-const theme = createTheme();
+vi.mock("@/utils/axiosInterceptor", () => ({
+  setAxiosRequestInterceptors: vi.fn(() => ({ type: "SET_INTERCEPTORS" })),
+}));
 
 describe("LoginButton", () => {
-  it("renders with the correct label", () => {
-    const store = createTestStore({});
+  const mockDispatch = vi.fn();
+  const theme = createTheme();
+
+  beforeEach(() => {
+    (useDispatch as unknown as Mock).mockReturnValue(mockDispatch);
+    mockDispatch.mockClear();
+  });
+
+  it("renders the button with the correct label", () => {
     render(
-      <Provider store={store}>
-        <ThemeProvider theme={theme}>
-          <LoginButton label="Sign In" />
-        </ThemeProvider>
-      </Provider>
+      <ThemeProvider theme={theme}>
+        <LoginButton label="Sign In" />
+      </ThemeProvider>
     );
 
     expect(screen.getByText("Sign In")).toBeInTheDocument();
   });
 
-  it("dispatches authenticate action on click", () => {
-    const store = createTestStore({});
-    store.dispatch = vi.fn();
-
+  it("dispatches authenticate and setAxiosRequestInterceptors on click", () => {
     render(
-      <Provider store={store}>
-        <ThemeProvider theme={theme}>
-          <LoginButton label="Login" />
-        </ThemeProvider>
-      </Provider>
+      <ThemeProvider theme={theme}>
+        <LoginButton label="Sign In" />
+      </ThemeProvider>
     );
 
-    const button = screen.getByRole("button");
-    fireEvent.click(button);
+    fireEvent.click(screen.getByRole("button", { name: /sign in/i }));
 
-    expect(store.dispatch).toHaveBeenCalledTimes(1);
-    expect(authenticate).toHaveBeenCalled();
+    expect(mockDispatch).toHaveBeenCalledWith(authenticate());
+    expect(mockDispatch).toHaveBeenCalledWith(setAxiosRequestInterceptors());
   });
 });
